requirements.txt
python_libnmap==0.7.0
python-docx==0.8.6
argparse==1.4.0
